% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tracks.R
\name{get_saved_tracks}
\alias{get_saved_tracks}
\title{get_saved_tracks}
\usage{
get_saved_tracks(select_key_cols = TRUE, rename_key_cols = TRUE)
}
\arguments{
\item{select_key_cols}{Boolean. Sets whether to select key columns of interest or alternatively to leave all columns as is. Defaults to TRUE}

\item{rename_key_cols}{Boolean. Sets whether to rename the key columns to more user-friendly names. Defaults to TRUE. Cannot be TRUE if \code{select_key_cols} is FALSE}
}
\value{
Dataframe
}
\description{
Creates a dataframe of all of the tracks you have saved (liked)
}
\examples{
\dontrun{
client_id = Sys.setenv("SPOTIFY_CLIENT_ID"),
client_secret = Sys.setenv("SPOTIFY_CLIENT_SECRET")
dat <- get_saved_tracks()
}
}
