% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spotifyr-functions.R
\name{get_spotify_authorization_code}
\alias{get_spotify_authorization_code}
\title{Get Spotify Authorization Code}
\usage{
get_spotify_authorization_code(
  client_id = Sys.getenv("SPOTIFY_CLIENT_ID"),
  client_secret = Sys.getenv("SPOTIFY_CLIENT_SECRET"),
  scope = get_scopes()
)
}
\arguments{
\item{client_id}{Defaults to System Envioronment variable "SPOTIFY_CLIENT_ID"}

\item{client_secret}{Defaults to System Envioronment variable "SPOTIFY_CLIENT_SECRET"}

\item{scope}{Space delimited string of spotify scopes,
found here: https://developer.spotify.com/documentation/general/guides/scopes/.
All scopes are selected by default}
}
\value{
The Spotify Web API Token2.0 reference class object (see
\code{httr::\link[httr]{oauth2.0_token}}), or an error message.
}
\description{
This function is copied verbatim from the \code{spotifyr} package by \href{https://www.rcharlie.com/spotifyr/index.html}{Charlie Thompson}. It is being used in accordance with its MIT License. All copyright is retained by C Thompson as follows
YEAR: 2017
COPYRIGHT HOLDER: Charlie Thompson
}
\details{
This function creates a Spotify authorization code.
See \code{httr::\link[httr]{oauth2.0_token}}.
}
\examples{
\dontrun{
authorization <- get_spotify_authorization_code()
}
}
\concept{authentication functions}
